+++ packages/eslint-plugin/rules/indent/indent._ts_.test.ts
@@ -867,6 +867,23 @@ run<RuleOptions, MessageIds>({
         | 'four'
       ;
     `,
+    // https://github.com/eslint-stylistic/eslint-stylistic/issues/909
+    $`
+      genericFunction<
+          () => void
+      >(
+          () => {
+              console.log("Test");
+          }
+      );
+    `,
+    $`
+      genericFunction<() => void>(
+          () => {
+              console.log("Test");
+          }
+      )
+    `,
     $`
       function foo<
           T